feat: add chains entrypoint
