--- packages/core/src/chains/index.ts
@@ -1,2 +1,0 @@
-export { configureChains } from './configureChains'
-export type { ConfigureChainsConfig } from './configureChains'
