+++ packages/core/src/constants/index.ts
@@ -3,13 +3,9 @@ export { erc20ABI, erc721ABI } from './abis'
 export { etherscanBlockExplorers } from './blockExplorers'
 export type { BlockExplorer, BlockExplorerName } from './blockExplorers'
 
-export {
-  chain,
-  chainId,
-  allChains,
-  defaultChains,
-  defaultL2Chains,
-} from './chains'
+export { chain, allChains, defaultChains, defaultL2Chains } from './chains'
+
+export { units } from './units'
 
 export {
   alchemyRpcUrls,
@@ -18,5 +14,3 @@ export {
   defaultInfuraId,
 } from './rpcs'
 export type { RpcProviderName } from './rpcs'
-
-export { units } from './units'
