+++ packages/core/src/constants/rpcs.ts
@@ -1,4 +1,4 @@
-import { ChainName } from './chains'
+import { ChainName } from '../chains'
 
 export const defaultAlchemyId = '_gg7wSSi0KMBsdKnGVfHDueq6xMB9EkC'
 export const defaultInfuraId = '84842078b09946638c03157f83405213'
