+++ packages/core/src/index.test.ts
@@ -34,7 +34,6 @@ it('should expose correct exports', () => {
       "watchSigner",
       "watchWebSocketProvider",
       "writeContract",
-      "configureChains",
       "createClient",
       "Client",
       "Connector",
@@ -62,6 +61,7 @@ it('should expose correct exports', () => {
       "UserRejectedRequestError",
       "createStorage",
       "noopStorage",
+      "configureChains",
       "normalizeChainId",
     ]
   `)
