+++ packages/core/src/index.ts
@@ -84,8 +84,7 @@ export type {
   WriteContractResult,
 } from './actions'
 
-export { configureChains } from './chains'
-export type { ConfigureChainsConfig } from './chains'
+export { chainId } from './chains'
 
 export { createClient, Client } from './client'
 export type { ClientConfig } from './client'
@@ -97,7 +96,6 @@ export {
   alchemyRpcUrls,
   allChains,
   chain,
-  chainId,
   defaultChains,
   defaultL2Chains,
   erc20ABI,
@@ -133,4 +131,5 @@ export type {
   WebSocketProvider,
 } from './types'
 
-export { normalizeChainId } from './utils'
+export { configureChains, normalizeChainId } from './utils'
+export type { ConfigureChainsConfig } from './utils'
