+++ packages/core/src/types/index.ts
@@ -8,20 +8,27 @@ import {
 } from '../constants'
 
 export type Chain = {
+  /** ID in number form */
   id: number
-  name: AddEthereumChainParameter['chainName']
+  /** Human-readable name */
+  name: string
+  /** ethers internal network name */
   network: string
+  /** Currency used by chain */
   nativeCurrency?: AddEthereumChainParameter['nativeCurrency']
+  /** Collection of RPC endpoints */
   rpcUrls: { [key in RpcProviderName]?: string } & {
     [key: string]: string
     default: string
   }
+  /** Collection of block explorers */
   blockExplorers?: {
-    [key in BlockExplorerName]: BlockExplorer
+    [key in BlockExplorerName]?: BlockExplorer
   } & {
     [key: string]: BlockExplorer
     default: BlockExplorer
   }
+  /** Flag for test networks */
   testnet?: boolean
 }
