+++ packages/core/src/utils/configureChains.test.ts
