+++ packages/core/src/utils/configureChains.ts
