+++ packages/core/src/utils/index.ts
@@ -1,3 +1,6 @@
+export { configureChains } from './configureChains'
+export type { ConfigureChainsConfig } from './configureChains'
+
 export { getInjectedName } from './getInjectedName'
 export { normalizeChainId } from './normalizeChainId'
 export { warn } from './warn'
