+++ packages/react/package.json
@@ -26,6 +26,10 @@
       "module": "./dist/wagmi.esm.js",
       "default": "./dist/wagmi.cjs.js"
     },
+    "./chains": {
+      "module": "./chains/dist/wagmi-chains.esm.js",
+      "default": "./chains/dist/wagmi-chains.cjs.js"
+    },
     "./connectors/coinbaseWallet": {
       "module": "./connectors/coinbaseWallet/dist/wagmi-connectors-coinbaseWallet.esm.js",
       "default": "./connectors/coinbaseWallet/dist/wagmi-connectors-coinbaseWallet.cjs.js"
@@ -46,10 +50,6 @@
       "module": "./providers/alchemy/dist/wagmi-providers-alchemy.esm.js",
       "default": "./providers/alchemy/dist/wagmi-providers-alchemy.cjs.js"
     },
-    "./providers/public": {
-      "module": "./providers/public/dist/wagmi-providers-public.esm.js",
-      "default": "./providers/public/dist/wagmi-providers-public.cjs.js"
-    },
     "./providers/infura": {
       "module": "./providers/infura/dist/wagmi-providers-infura.esm.js",
       "default": "./providers/infura/dist/wagmi-providers-infura.cjs.js"
@@ -57,9 +57,14 @@
     "./providers/jsonRpc": {
       "module": "./providers/jsonRpc/dist/wagmi-providers-jsonRpc.esm.js",
       "default": "./providers/jsonRpc/dist/wagmi-providers-jsonRpc.cjs.js"
+    },
+    "./providers/public": {
+      "module": "./providers/public/dist/wagmi-providers-public.esm.js",
+      "default": "./providers/public/dist/wagmi-providers-public.cjs.js"
     }
   },
   "files": [
+    "/chains",
     "/connectors",
     "/dist",
     "/providers"
@@ -67,6 +72,7 @@
   "preconstruct": {
     "entrypoints": [
       "index.ts",
+      "chains.ts",
       "connectors/coinbaseWallet.ts",
       "connectors/injected.ts",
       "connectors/metaMask.ts",
