+++ packages/react/src/chains.ts
@@ -1,0 +1,16 @@
+export {
+  arbitrum,
+  arbitrumRinkeby,
+  avalanche,
+  goerli,
+  hardhat,
+  kovan,
+  localhost,
+  mainnet,
+  optimism,
+  optimismKovan,
+  polygon,
+  polygonMumbai,
+  rinkeby,
+  ropsten,
+} from '@wagmi/core/chains'
